cdc/server/integration_test.go: added in a new test for the CDC server using the workload.Checker
Integrated the logic from the existing CDC server integration test and the
latest developments for the pglogical integration tests to create a test based
on the workload.Checker that validates the CDC server. There are variants of
this test that support various changefeed configuration options such as
key_in_value == diff == queries == true or setting webhook == false, among others.
This gives confidence that various ways that customers define changefeeds will
work with the replicator CDC server.

There are now helper methods that help to create changefeeds for a variety
of configurations, along with relevant tests that verify the create changefeed
string is as expected. There are also new helpers added to help gate when the
workload checker runs since there is a minimum version requirement for its
operation. These shared utilities now live within the sinktest package.

Resolves: #906
Release Note: None