cdcjson/query_payload: updated the unmarshal JSON logic so that it uses the key values from the CDC message if it's provided; otherwise, it does the back-processing to obtain the key values
Previously, query payloads without key_in_value may not have the information
necessary to do so when processing deletes, when the target PK requires a
projected value. This change supports messages that come in with the
key_in_value data. This obviates back-parsing if it's explicitly provided.
Otherwise, we fallback and use the back-parsing in the case where this data is
not provided.

This change should help unblock customers we are currently blocked by
performing deletes in certain circumstances.

There is also now a set of tests at the unit and integration level to ensure
that the key_in_value case and otherwise are handled.

Resolves: #1044
Release Note: None