mylogical: added in handling for unsigned bigints
Previously, bigint unsigned ould be interpreted as int64, which meant that over-large values would overflow and be interpreted as negative values. This change aims to make mylogical aware of signedness and treat the bigint unsigned case as uint64, which means that over-large values can now be interpreted properly. There is also logic introduced to get the signedness from the schema for older MySQL dialects.

Resolves: CC-30721
Release Note: Fixing issue with unsigned bigints overflowing.